refactor: Move to operations directory.
